refactor: 멀티모듈 전환으로 인한 import 및 코드 변경
