+++ interface-module/src/main/java/org/c4marathon/assignment/usecase/member/MemberAccountUseCase.java
@@ -1,6 +1,8 @@
 package org.c4marathon.assignment.usecase.member;
 
+import org.c4marathon.assignment.domain.model.MainAccount;
 import org.c4marathon.assignment.domain.model.Member;
+import org.c4marathon.assignment.domain.service.MainAccountService;
 import org.c4marathon.assignment.domain.service.MemberService;
 import org.c4marathon.assignment.api.member.dto.MemberRegistrationRequestDto;
 import org.springframework.transaction.annotation.Transactional;
@@ -16,7 +18,7 @@ public class MemberAccountUseCase {
 	private final MainAccountService mainAccountService;
 
 	@Transactional
-	public String registerMemberWithAccount(MemberRegistrationRequestDto request) {
+	public MainAccount registerMemberWithAccount(MemberRegistrationRequestDto request) {
 		Member member = Member.builder()
 			.name(request.name())
 			.email(request.email())
