+++ reminder-module/src/main/java/org/c4marathon/assignment/domain/service/ReminderService.java
@@ -3,8 +3,9 @@ package org.c4marathon.assignment.domain.service;
 import java.util.List;
 import java.util.Map;
 
-import org.c4marathon.assignment.domain.model.member.Member;
-import org.c4marathon.assignment.domain.model.PendingTransfer;import org.springframework.stereotype.Service;
+import org.c4marathon.assignment.domain.model.Member;
+import org.c4marathon.assignment.domain.model.PendingTransfer;
+import org.springframework.stereotype.Service;
 
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -14,14 +15,14 @@ import lombok.extern.slf4j.Slf4j;
 @AllArgsConstructor
 public class ReminderService {
 
-	public void remindTransactions(PendingTransfer transferTransaction) {
-		log.info("[Remind] Transaction ID: {}, Amount: {}, ExpiredAt: {}, ToMember: {}",
-			transferTransaction.getId(),
-			transferTransaction.getAmount(),
-			transferTransaction.getExpiredAt(),
-			transferTransaction.getToMainAccount().getMember().getName()
-		);
-	}
+	// public void remindTransactions(PendingTransfer transferTransaction) {
+	// 	log.info("[Remind] Transaction ID: {}, Amount: {}, ExpiredAt: {}, ToMember: {}",
+	// 		transferTransaction.getId(),
+	// 		transferTransaction.getAmount(),
+	// 		transferTransaction.getExpiredAt(),
+	// 		transferTransaction.getToMainAccount().getMember().getName()
+	// 	);
+	// }
 
 	public void remindTransactions(Map<Member, List<PendingTransfer>> grouped) {
 		grouped.forEach((member, transactions) -> {
